Refactor: admin 패키지 이름 오타 수정
