+++ src/main/java/shop/kkeujeok/kkeujeokbackend/admin/domain/repository/NoticeRepository.java
@@ -1,7 +1,7 @@
-package shop.kkeujeok.kkeujeokbackend.amdin.domain.repository;
+package shop.kkeujeok.kkeujeokbackend.admin.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.kkeujeok.kkeujeokbackend.amdin.domain.Notice;
+import shop.kkeujeok.kkeujeokbackend.admin.domain.Notice;
 
 public interface NoticeRepository extends JpaRepository<Notice, Long> {
 }
